==> 4. Regression/3. Selection.Rmd <==
# Selección de atributos

## Imports necesarios

```{r}
#install.packages("ggplot2")
#install.packages("dplyr")
#install.packages("tidyr")
#install.packages("randomForest")
#install.packages("scales")
```

```{r}
library(ggplot2)
library(dplyr)
library(tidyr)
library(randomForest)
library(scales)
```

## 1. Funciones auxiliares

### 1.1. Reducción de la dimensionalidad generada por One-Hot encoding haciendo uso de PCA

En esta sección, se implementa la técnica de Análisis de Componentes Principales (PCA, por sus siglas en inglés) para reducir la dimensionalidad de las variables después de aplicar One-Hot Encoding. La dimensionalidad reducida facilita el manejo eficiente de los datos, disminuyendo la complejidad y el costo computacional en modelos de machine learning.

```{r}

reduce_dimensionality <- function(data, patterns, new_names) {
  if (length(patterns) != length(new_names)) {
    stop("Las listas de patrones y nombres nuevos deben tener la misma longitud")
  }

  for (i in seq_along(patterns)) {
    columns <- grepl(patterns[i], colnames(data))
    data_columns <- data[, columns]

    # Manejar datos faltantes
    data_columns <- na.omit(data_columns)

    # Escalar los datos
    data_columns <- scale(data_columns)

    # Realizar PCA con prcomp en lugar de princomp
    pca_result <- prcomp(data_columns, scale. = TRUE)

    # Seleccionar los primeros N componentes principales
    N <- min(ncol(data_columns), 3)  # Limitar a un máximo de 3 componentes
    pca_columns <- pca_result$x[, 1:N]

    # Normalizar los valores de PCA entre 0 y 1
    pca_columns <- rescale(pca_columns)
    
    new_column_names <- paste0(new_names[i], ".PCA", 1:N)

    data <- data[, !columns]
    data <- cbind(data, setNames(as.data.frame(pca_columns), new_column_names))
  }

  return(data)
}

```

```{r}

calcular_indices <- function(valores, pattern){
  
  # Crear un vector vacío para almacenar los índices
  indices <- numeric(length(valores))
  
  # Iterar sobre cada valor
  for (i in seq_along(valores)) {
    # Obtener el índice de la primera fila que tiene un 1 para este valor
    indices[i] <- which(data[[valores[i]]] == 1)[1]
  }
  
  # Crear un dataframe con los valores y sus índices correspondientes
  df_indices <- data.frame(Variable = valores, Indice = indices, Pattern = pattern)
  
  return (df_indices)
}
```

## 2. Aplicación del proceso de selección de atributos

Lectura de los datos obtenidos tras aplicar el proceso de transformación.

```{r}

data <- read.csv("../data/CleanData/datos_transformados.csv")
head(data)
```

Eliminar aquellas columnas que no aportan información al problema.

La razón detrás de la eliminación de las columnas "Número de Estudiantes", "Curso" y "Universidad" se basa en consideraciones específicas que afectan la calidad y la relevancia de los datos para el análisis y modelización

1.  **Número de Estudiantes:**

    -   La columna "Número de Estudiantes" fue eliminada debido a la presencia de valores iguales a cero cuando la zona de nacionalidad era diferente a España. Para lograr una mayor generalización y consideración de nacionalidades diversas, la eliminación de esta columna resultó más adecuada, permitiendo una representación más completa de los datos en relación con las diferentes nacionalidades.

2.  **Curso:**

    -   La columna "Curso" fue eliminada al decidir no tener en cuenta análisis de series temporales en el estudio. Al excluir esta variable temporal, se simplifica el conjunto de datos, enfocándolo en aspectos estáticos y facilitando la interpretación del modelo sin consideraciones temporales.

3.  **Sexo:**

    -   La columna "Sexo" fue eliminada para evitar introducir sesgos en los datos, de manera que las predicciones realizadas por los modelos no se vean afectadas por el valor de dicha columna.

4.  **Universidad:**

    -   La columna "Universidad" fue eliminada debido a la limitada cantidad de datos disponibles para cada universidad específica. En lugar de modelar por universidad, se optó por generalizar a nivel de comunidad autónoma para asegurar una representación más robusta y evitar posibles sesgos causados por la escasez de datos por universidad.

```{r}

data <- data[, -which(names(data) == "Número.de.estudiantes")]
data <- data[, -which(names(data) == "Curso")]
data <- data[, -which(names(data) == "Sexo")]
data <- data %>% select(-starts_with("Universidad"))
```

```{r}

comunidades <- c("Comunidad.AutónomaAndalucía","Comunidad.AutónomaAragón", "Comunidad.AutónomaAsturias", "Comunidad.AutónomaCantabria", "Comunidad.AutónomaCastilla.y.León", "Comunidad.AutónomaCataluña", "Comunidad.AutónomaExtremadura", "Comunidad.AutónomaGalicia", "Comunidad.AutónomaLa.Rioja", "Comunidad.AutónomaMadrid", "Comunidad.AutónomaMurcia", "Comunidad.AutónomaNavarra", "Comunidad.AutónomaPaís.Vasco")

ramas_enseñanza <- c("Rama.de.enseñanzaCiencias", "Rama.de.enseñanzaCiencias.de.la.Salud", "Rama.de.enseñanzaCiencias.Sociales.y.Jurídicas", "Rama.de.enseñanzaIngeniería.y.Arquitectura")

ambitos_estudio <- c("Ámbito.de.estudioArquitectura.y.construcción", "Ámbito.de.estudioArtes", "Ámbito.de.estudioCiencias.de.la.vida", "Ámbito.de.estudioCiencias.Físicas..químicas..geológicas", "Ámbito.de.estudioDerecho", "Ámbito.de.estudioEconomía", "Ámbito.de.estudioEnfermería.y.atención.a.enfermos", "Ámbito.de.estudioHumanidades", "Ámbito.de.estudioInformática", "Ámbito.de.estudioIngenierías", "Ámbito.de.estudioLenguas", "Ámbito.de.estudioMedicina", "Ámbito.de.estudioPeriodismo.e.información", "Ámbito.de.estudioTrabajo.social.y.orientación")

zonas_nacionalidad <- c("Zona.de.nacionalidadAsia.y.Oceanía", "Zona.de.nacionalidadEEUU.y.Canadá", "Zona.de.nacionalidadEspaña", "Zona.de.nacionalidadNorte.de.África", "Zona.de.nacionalidadResto.de.África", "Zona.de.nacionalidadResto.de.Europa", "Zona.de.nacionalidadUnión.Europea")

df_comunidades <- calcular_indices(comunidades, "^Comunidad.PCA")
df_ramas_enseñanza <- calcular_indices(ramas_enseñanza, "^Rama.PCA")
df_ambitos_estudio <- calcular_indices(ambitos_estudio, "^Ámbito.PCA")
df_zonas_nacionalidad <- calcular_indices(zonas_nacionalidad, "^Zona.de.nacionalidad.PCA")

# Unificar todos los dataframes en uno solo
df_indices <- rbind(df_comunidades, df_ramas_enseñanza, df_ambitos_estudio, df_zonas_nacionalidad)

head(df_indices)
```

Reducir la dimensionalidad generada tras haber realizado el proceso de One-Hot encoding.

```{r}
patterns <- c("^Comunidad", "^Rama", "^Ámbito", "^Zona.de.nacionalidad")
new_names <- c("Comunidad", "Rama", "Ámbito", "Zona.de.nacionalidad")
df_res <- reduce_dimensionality(data, patterns, new_names)
head(df_res)
```

```{r}

# Función para obtener las PCA según un patrón
get_pca_values <- function(df, df_indices) {
  df_pca <- data.frame()
  variables <- df_indices$Variable
  indices <- df_indices$Indice
  patterns <- df_indices$Pattern
  
  for (i in 1:length(indices)) {
    # Obtener el índice y el nombre de la comunidad de la fila actual
    variable <- variables[i]
    indice <- as.numeric(indices[i])
    pattern <- patterns[i]
    
    # Obtener las columnas que comienzan con el patrón
    pca_columns <- grepl(pattern, colnames(df))
    
    # Obtener los valores de esas columnas para el índice especificado en indices
    matching_rows <- df[indice, ]
    
    # Verificar si hay coincidencias antes de proceder
    if (nrow(matching_rows) > 0) {
      pca_values <- matching_rows[, pca_columns]
      # Crear un dataframe temporal con los valores de PCA y el nombre de la variable
      colnames(pca_values) <- paste0("PCA", seq_len(ncol(pca_values)))

      df_temp <- as.data.frame(t(pca_values))
      
      df_temp <- as.data.frame(t(df_temp))
      
      df_temp$Variable <- variable
      
      # Añadir el dataframe temporal al dataframe df_pca
      df_pca <- rbind(df_pca, df_temp)
      
      row.names(df_pca)[i] <- i
    } else {
      cat("No se encontraron coincidencias para el índice:", indice, "\n")
    }
  }
  
  return(df_pca)
}
```

Almacenar en un dataframe todos los componentes de PCA generados, para, en el archivo Model.Rmd, poder determinar automáticamente transformación entre las variables categóricas intruducidas por el usuario, como Comunidad Autónoma, y sus PCAs asociadas.

```{r}

df_PCAs = get_pca_values(df_res, df_indices)
print(df_PCAs)
```

Redondear todos los atributos numéricos a dos decimales.

```{r}
# Redondear a 2 decimales todos los atributos numéricos
df_res <- df_res %>% mutate_if(is.numeric, round, 2)
df_PCAs <- df_PCAs %>% mutate_if(is.numeric, round, 2)
```

Uso de Random Forest para ver la importancia de las distintas características con la columna "Nota media".

```{r}
set.seed(1234)
rf_model <- randomForest(Nota.media ~ ., data = df_res, ntree = 10, importance = TRUE)

print(importance(rf_model))
```

```{r}
# Gráfico de la importancia de los atributos
varImpPlot(rf_model)
```

## 5. Exportación de los datos resultantes a un fichero CSV

```{r}

nombre_archivo <- "../data/CleanData/datos_seleccion_atributos.csv"
write.csv(df_res, nombre_archivo, row.names = FALSE)
cat("Datos exportados a", nombre_archivo, "\n")
```

```{r}

nombre_archivo <- "../data/CleanData/var_PCAs.csv"
write.csv(df_PCAs, nombre_archivo, row.names = FALSE)
cat("Datos de PCAs exportados a", nombre_archivo, "\n")
```
